rename SpaceViewClass(Impl)
